tests/wpt/web-platform-tests/credential-management/fedcm-network-requests.sub.https.html

<!DOCTYPE html>
<title>Federated Credential Management API network request tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>

<script type="module">
import {set_fedcm_cookie} from './support/fedcm-helper.js';

const url_prefix = 'https://{{host}}:{{ports[https][0]}}/credential-management/support/';
promise_test(async t => {
  await set_fedcm_cookie();
  const result = await navigator.credentials.get({
    federated: {
      providers: [{
        url: url_prefix,
        clientId: '1',
      }]
    }
  });
  const token = await result.login({nonce: '1'});
  assert_equals(token.idToken, 'token');
}, 'Test that COEP policy do not apply to FedCM requests');
</script>
